Upgraded to Hibernate 4.1 using Spring 4's Hibernate integration classes.

+++ opennms-webapp-rest/src/main/java/org/opennms/web/rest/v1/AlarmRestService.java
@@ -109,7 +109,7 @@ public class AlarmRestService extends AlarmRestServiceBase {
     @Transactional
     public String getCount(@Context SecurityContext securityContext) {
         assertUserReadCredentials(securityContext);
-        return Integer.toString(m_alarmDao.countAll());
+        return Long.toString(m_alarmDao.countAll());
     }
 
     /**
