+++ opennms-webapp-rest/src/main/java/org/opennms/web/rest/v2/AbstractDaoRestService.java
@@ -146,7 +146,7 @@ public abstract class AbstractDaoRestService<T,K extends Serializable> {
 			crit.setLimit(null);
 			crit.setOffset(null);
 			crit.setOrders(new ArrayList<Order>());
-			int totalCount = getDao().countMatching(crit);
+			long totalCount = getDao().countMatching(crit);
 
 			JaxbListWrapper<T> list = createListWrapper(coll);
 			list.setTotalCount(totalCount);
